feat: update the about text
