+++ pages/about/styles.module.scss
@@ -18,7 +18,7 @@
     text-align: justify;
     font-size: 1.1rem;
     text-indent: 4rem;
-    margin: 0.5rem 1.5rem;
+    margin: 1rem 1.5rem;
   }
 }
